+++ trunk/source/newtonian/two_dimensional/point_motions/lagrangian.cpp
@@ -4,5 +4,5 @@ Vector2D Lagrangian::CalcVelocity
 (int index, Tessellation const& /*tessellation*/,
  vector<Primitive> const& primitives,double /*time*/)
 {
-  return primitives[index].Velocity;
+  return primitives[(size_t)index].Velocity;
 }
