+++ trunk/source/tessellation/voronoi_logger.cpp
@@ -56,7 +56,7 @@ void BinLogger::output(Tessellation const& v)
 		const vector<int> indices = v.GetCellEdges(i);
 		binary_write_single_int((int)indices.size(),file_handle);
 		for(int j=0;j<(int)indices.size();++j)
-			binary_write_single_int(indices[j],file_handle);
+		  binary_write_single_int(indices[(size_t)j],file_handle);
 	}
 	file_handle.close();
 }
@@ -101,7 +101,7 @@ void BinLogger::output(VoronoiMesh const& v)
 		const vector<int> indices = v.GetCellEdges(i);
 		binary_write_single_int((int)indices.size(),file_handle);
 		for(int j=0;j<(int)indices.size();++j)
-			binary_write_single_int(indices[j],file_handle);
+		  binary_write_single_int(indices[(size_t)j],file_handle);
 	}
 	file_handle.close();
 }
